requirements.txt
boto3>=1.34.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
streamlit>=1.29.0
pandas>=2.0.0
langchain>=0.1.0
langchain-aws>=0.1.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
